uar-495 move tests and fix code smells
